Refactor Docs View
